requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
SQLAlchemy
openai
python-dotenv
ollama
Flask-Markdown
